+++ docs/.vitepress/theme/components/WobbleMaterialDemo.vue
@@ -1,0 +1,27 @@
+<script setup lang="ts">
+import { TresCanvas } from '@tresjs/core'
+import { OrbitControls, MeshWobbleMaterial } from '@tresjs/cientos'
+import { BasicShadowMap, SRGBColorSpace, NoToneMapping } from 'three'
+
+const gl = {
+  clearColor: '#82DBC5',
+  shadows: true,
+  alpha: false,
+  shadowMapType: BasicShadowMap,
+  outputColorSpace: SRGBColorSpace,
+  toneMapping: NoToneMapping,
+}
+</script>
+
+<template>
+  <TresCanvas v-bind="gl" ref="context">
+    <TresPerspectiveCamera :position="[3, 3, 3]" />
+    <TresMesh>
+      <TresTorusGeometry />
+      <MeshWobbleMaterial color="orange" :speed="1" :factor="2" />
+    </TresMesh>
+    <TresAmbientLight :intensity="1" />
+    <TresDirectionalLight :intensity="1" :position="[2, 2, 2]" />
+    <OrbitControls />
+  </TresCanvas>
+</template>