+++ docs/components.d.ts
@@ -14,5 +14,6 @@ declare module 'vue' {
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     WobbleMaterial: typeof import('./.vitepress/theme/components/WobbleMaterial.vue')['default']
+    WobbleMaterialDemo: typeof import('./.vitepress/theme/components/WobbleMaterialDemo.vue')['default']
   }
 }