+++ playground/src/pages/BackdropDemo.vue
@@ -16,10 +16,10 @@ const gl = {
 
 <template>
   <TresCanvas v-bind="gl">
-    <TresPerspectiveCamera :position="[3, 3, 3]" />
+    <TresPerspectiveCamera :position="[11, 11, 11]" />
     <OrbitControls />
     <Backdrop :floor="1.5" :scale="[20, 5, 5]">
-        <TresMeshPhysicalMaterial  :roughness="1" color="#efefef" :side="2" />
+      <TresMeshPhysicalMaterial :roughness="1" color="#efefef" :side="2" />
     </Backdrop>
     <TresAmbientLight :intensity="1" />
   </TresCanvas>