+++ playground/src/pages/FirstPersonControlsDemo.vue
@@ -1,14 +1,7 @@
 <script setup lang="ts">
-import { ref, watch, watchEffect } from 'vue'
 import { TresCanvas } from '@tresjs/core'
 import { BasicShadowMap, NoToneMapping } from 'three'
 import { PointerLockControls, KeyboardControls } from '@tresjs/cientos'
-
-const PLControlsRef = ref<PointerLockControls | null>(null)
-
-watch(PLControlsRef, value => {
-  console.log(value)
-})
 
 const gl = {
   clearColor: '#82DBC5',
@@ -24,7 +17,7 @@ const isActive = (state: boolean) => console.log(state)
 <template>
   <TresCanvas v-bind="gl">
     <TresPerspectiveCamera :position="[0, 3, 10]" />
-    <PointerLockControls ref="PLControlsRef" make-default @is-lock="state => isActive(state)" />
+    <PointerLockControls make-default @is-lock="state => isActive(state)" />
     <KeyboardControls head-bobbing />
 
     <TresGridHelper :args="[100, 100]" />