+++ playground/src/router.ts
@@ -12,6 +12,11 @@ const routes = [
     component: () => import('./pages/OrbitControlsDemo.vue'),
   },
   {
+    path: '/controls/camera-controls',
+    name: 'CameraControls',
+    component: () => import('./pages/CameraControlsDemo.vue'),
+  },
+  {
     path: '/controls/transform-controls',
     name: 'TransformControls',
     component: () => import('./pages/TransformControlsDemo.vue'),
@@ -22,9 +27,19 @@ const routes = [
     component: () => import('./pages/FirstPersonControlsDemo.vue'),
   },
   {
+    path: '/controls/scroll-controls',
+    name: 'ScrollControls',
+    component: () => import('./pages/ScrollControlsDemo.vue'),
+  },
+  {
     path: '/abstractions/smoke',
     name: 'Smoke',
     component: () => import('./pages/SmokeDemo.vue'),
+  },
+  {
+    path: '/abstractions/levioso',
+    name: 'Levioso',
+    component: () => import('./pages/LeviosoDemo.vue'),
   },
   {
     path: '/abstractions/precipitation',
@@ -40,7 +55,27 @@ const routes = [
     path: '/abstractions/text-3d',
     name: 'Text3D',
     component: () => import('./pages/Text3DDemo.vue'),
-  }
+  },
+  {
+    path: '/abstractions/environment',
+    name: 'Environment',
+    component: () => import('./pages/EnvironmentDemo.vue'),
+  },
+  {
+    path: '/staging/backdrop',
+    name: 'Backdrop',
+    component: () => import('./pages/BackdropDemo.vue'),
+  },
+  {
+    path: '/abstractions/contact-shadows',
+    name: 'ContactShadows',
+    component: () => import('./pages/ContactShadowsDemo.vue'),
+  },
+  {
+    path: '/abstractions/mouse-parallax',
+    name: 'MouseParallax',
+    component: () => import('./pages/MouseParallaxDemo.vue'),
+  },
 ]
 
 export const router = createRouter({