+++ src/core/controls/OrbitControls.vue
@@ -2,10 +2,8 @@
 import { Camera, TOUCH } from 'three'
 import { OrbitControls } from 'three-stdlib'
 import { ref, watch, onUnmounted, toRefs } from 'vue'
-import { TresVector3, useRenderLoop } from '@tresjs/core'
+import { TresVector3, useRenderLoop, useTresContext } from '@tresjs/core'
 import { useEventListener } from '@vueuse/core'
-
-import { useCientos } from '../useCientos'
 
 export interface OrbitControlsProps {
   /**
@@ -278,48 +276,48 @@ const {
   target,
 } = toRefs(props)
 
-const { state, setState, extend } = useCientos()
+const { camera: activeCamera, renderer, extend, controls } = useTresContext()
 
-const controls = ref<OrbitControls | null>(null)
+const controlsRef = ref<OrbitControls | null>(null)
 
 extend({ OrbitControls })
 
-watch(controls, value => {
+watch(controlsRef, value => {
   addEventListeners()
   if (value && makeDefault.value) {
-    setState('controls', value)
+    controls.value = value
   } else {
-    setState('controls', null)
+    controls.value = null
   }
 })
 
 const emit = defineEmits(['change', 'start', 'end'])
 
 function addEventListeners() {
-  useEventListener(controls.value as any, 'change', () => emit('change', controls.value))
-  useEventListener(controls.value as any, 'start', () => emit('start', controls.value))
-  useEventListener(controls.value as any, 'end', () => emit('end', controls.value))
+  useEventListener(controlsRef.value as any, 'change', () => emit('change', controlsRef.value))
+  useEventListener(controlsRef.value as any, 'start', () => emit('start', controlsRef.value))
+  useEventListener(controlsRef.value as any, 'end', () => emit('end', controlsRef.value))
 }
 
 const { onLoop } = useRenderLoop()
 
 onLoop(() => {
-  if (controls.value && (enableDamping.value || autoRotate.value)) {
-    controls.value.update()
+  if (controlsRef.value && (enableDamping.value || autoRotate.value)) {
+    controlsRef.value.update()
   }
 })
 
 onUnmounted(() => {
-  if (controls.value) {
-    controls.value.dispose()
+  if (controlsRef.value) {
+    controlsRef.value.dispose()
   }
 })
 </script>
 
 <template>
   <TresOrbitControls
-    v-if="state.camera && state.renderer"
-    ref="controls"
+    v-if="activeCamera && renderer"
+    ref="controlsRef"
     :target="target"
     :auto-rotate="autoRotate"
     :auto-rotate-speed="autoRotateSpeed"
@@ -341,6 +339,6 @@ onUnmounted(() => {
     :zoom-speed="zoomSpeed"
     :enable-rotate="enableRotate"
     :rotate-speed="rotateSpeed"
-    :args="[state.camera || camera, state.renderer?.domElement || domElement]"
+    :args="[activeCamera || camera, renderer?.domElement || domElement]"
   />
 </template>